+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Don't throw an error when a `al-field` element has no field name configured
 * Overwrite preloaded `al-select` values with their processed variant
 * Throw an error when a document does not have a `loadQueryBuilderData` method while using the `FormApi#queryBuilderData` action
+* `QueryBuilderValue` instances will now always add a type to the `value_explicit` object
 
 ## 0.2.0 (2022-11-02)